+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## unreleased
+* Add plan create/update/find API endpoint
 * Add `TransactionReview` webhook notification
 
 ## 6.4.1